Change uint32 ms types to double
We are dealing w/ floats here so the metric type needs to be updated
accordingly.

Closes #106

(cherry picked from commit b50a91b132b41b07457095d28516f26358b7dc1a)

Conflicts:
	README.md

+++ glance_api_local_check.py
@@ -66,7 +66,7 @@ def check(auth_ref, args):
     # only want to send other metrics if api is up
     if is_up:
         metric('glance_api_local_response_time',
-               'uint32',
+               'double',
                '%.3f' % milliseconds,
                'ms')
         for status in IMAGE_STATUSES:
